@@ -7,7 +7,7 @@
   "config": {
     "autoloader-suffix": "AdminColumns",
     "platform": {
-      "php": "7.2.0"
+      "php": "7.4"
     }
   },
   "require": {
