@@ -2,7 +2,7 @@
 
 namespace AC\Admin;
 
-use AC\Asset\Location\Absolute;
+use AC\AdminColumns;
 use AC\Registerable;
 
 class Admin implements Registerable
@@ -16,10 +16,10 @@ class Admin implements Registerable
 
     private $scripts;
 
-    public function __construct(RequestHandlerInterface $request_handler, Absolute $location, AdminScripts $scripts)
+    public function __construct(RequestHandlerInterface $request_handler, AdminColumns $plugin, AdminScripts $scripts)
     {
         $this->request_handler = $request_handler;
-        $this->location = $location;
+        $this->location = $plugin->get_location();
         $this->scripts = $scripts;
     }
 
