@@ -6,20 +6,20 @@ use AC\Admin;
 use AC\Admin\MenuFactoryInterface;
 use AC\Admin\Page;
 use AC\Admin\PageFactoryInterface;
-use AC\Asset\Location;
 use AC\Deprecated\HookCollectionFactory;
 use AC\Deprecated\Hooks;
+use AC\Entity\Plugin;
 
 class Help implements PageFactoryInterface
 {
 
-    protected Location\Absolute $location;
+    protected Plugin $plugin;
 
     protected MenuFactoryInterface $menu_factory;
 
-    public function __construct(Location\Absolute $location, MenuFactoryInterface $menu_factory)
+    public function __construct(Plugin $plugin, MenuFactoryInterface $menu_factory)
     {
-        $this->location = $location;
+        $this->plugin = $plugin;
         $this->menu_factory = $menu_factory;
     }
 
@@ -27,7 +27,7 @@ class Help implements PageFactoryInterface
     {
         return new Page\Help(
             new Hooks(new HookCollectionFactory()),
-            $this->location,
+            $this->plugin,
             new Admin\View\Menu($this->menu_factory->create('help'))
         );
     }
