@@ -4,28 +4,28 @@ declare(strict_types=1);
 
 namespace AC;
 
-use AC\Asset\Location\Absolute;
+use AC\Asset\Location;
 use AC\ListScreenRepository\Storage;
-use AC\Vendor\Psr\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 use LogicException;
 
-class Container
+final class Container
 {
 
-    private static $instance;
+    private static ?ContainerInterface $instance = null;
 
     public static function set_container(ContainerInterface $container): void
     {
-        if (self::$instance) {
+        if (self::$instance !== null) {
             throw new LogicException('Container is already set.');
         }
 
         self::$instance = $container;
     }
 
-    public static function get_location(): Absolute
+    public static function get_location(): Location
     {
-        return self::$instance->get(Absolute::class);
+        return self::$instance->get(AdminColumns::class)->get_location();
     }
 
     public static function get_storage(): Storage
